SCP-3192: Refactored modules
