+++ marlowe-actus/marlowe-actus.cabal
@@ -34,10 +34,10 @@ source-repository head
   location: https://github.com/input-output-hk/plutus
 
 flag defer-plugin-errors
-    description:
-        Defer errors from the plugin, useful for things like Haddock that can't handle it.
-    default: False
-    manual: True
+  description:
+      Defer errors from the plugin, useful for things like Haddock that can't handle it.
+  default: False
+  manual: True
 
 library actus-core
   hs-source-dirs: actus-core/src
@@ -75,103 +75,30 @@ library actus-core
     -Wincomplete-uni-patterns -Wincomplete-record-updates
     -Wredundant-constraints -Widentities
 
-library actus-haskell
-  hs-source-dirs: actus-haskell/src
-  build-depends:
-    aeson -any,
-    base -any,
-    actus-core,
-    bytestring,
-    containers -any,
-    newtype-generics,
-    template-haskell -any,
-    text,
-    vector,
-    mtl -any,
-    time -any,
-    sort -any,
-    validation -any
+test-suite actus-core-test
   default-language: Haskell2010
-  default-extensions: ExplicitForAll ScopedTypeVariables
-                      DeriveGeneric StandaloneDeriving DeriveLift
-                      DeriveFunctor DeriveFoldable
-                      DeriveTraversable
-  exposed-modules:
-    Actus.Haskell
-  ghc-options:
-    -Wall -Wnoncanonical-monad-instances -Wunused-packages
-    -Wincomplete-uni-patterns -Wincomplete-record-updates
-    -Wredundant-constraints -Widentities
-
-library actus-testframework
-    default-language: Haskell2010
-    hs-source-dirs: actus-haskell/test
-    type: exitcode-stdio-1.0
-    exposed-modules:
-        Spec.Actus.TestFramework
-    build-depends:
-        actus-core,
-        actus-haskell,
-        vector -any,
-        unordered-containers -any,
-        scientific -any,
-        extra -any,
-        utf8-string -any,
-        aeson -any,
-        base >=4.9 && <5,
-        containers -any,
-        hedgehog -any,
-        lens -any,
-        bytestring -any,
-        tasty -any,
-        tasty-hunit -any,
-        tasty-quickcheck -any,
-        text -any,
-        cborg,
-        template-haskell -any,
-        time -any,
-        validation -any,
-        pretty-simple -any,
-        mtl -any,
-        sort -any,
-        QuickCheck
-  ghc-options:
-    -Wall -Wnoncanonical-monad-instances
-    -Wincomplete-uni-patterns -Wincomplete-record-updates
-    -Wredundant-constraints -Widentities
-
-test-suite actus-haskell-test
-    default-language: Haskell2010
-    hs-source-dirs: actus-haskell/test
-    type: exitcode-stdio-1.0
-    main-is: Spec.hs
-    build-depends:
-        actus-core,
-        actus-haskell,
-        actus-testframework,
-        vector -any,
-        unordered-containers -any,
-        scientific -any,
-        extra -any,
-        utf8-string -any,
-        aeson -any,
-        base >=4.9 && <5,
-        containers -any,
-        hedgehog -any,
-        lens -any,
-        bytestring -any,
-        tasty -any,
-        tasty-hunit -any,
-        tasty-quickcheck -any,
-        text -any,
-        cborg,
-        template-haskell -any,
-        time -any,
-        validation -any,
-        pretty-simple -any,
-        mtl -any,
-        sort -any,
-        QuickCheck
+  hs-source-dirs: actus-core/test
+  type: exitcode-stdio-1.0
+  main-is: Spec.hs
+  other-modules:
+      Spec.Actus.Haskell
+      Spec.Actus.TestFramework
+  build-depends:
+    actus-core,
+    aeson -any,
+    base >=4.9 && <5,
+    bytestring -any,
+    containers -any,
+    lens -any,
+    mtl -any,
+    pretty-simple -any,
+    sort -any,
+    tasty -any,
+    tasty-hunit -any,
+    tasty-quickcheck -any,
+    text -any,
+    time -any,
+    QuickCheck
   ghc-options:
     -Wall -Wnoncanonical-monad-instances
     -Wincomplete-uni-patterns -Wincomplete-record-updates
@@ -181,18 +108,12 @@ library
   hs-source-dirs: src
   build-depends:
     actus-core,
-    actus-haskell,
-    aeson -any,
     base -any,
     bytestring,
-    containers -any,
-    newtype-generics,
-    template-haskell -any,
     plutus-tx -any,
     plutus-contract -any,
     plutus-ledger,
     text,
-    vector,
     marlowe -any,
     mtl -any,
     time -any,
@@ -212,48 +133,41 @@ library
     -Wredundant-constraints -Widentities
 
 test-suite marlowe-actus-test
-    default-language: Haskell2010
-    hs-source-dirs: test
-    type: exitcode-stdio-1.0
-    main-is: Spec.hs
-    other-modules:
-        Spec.Marlowe.ACTUS.Examples
-        Spec.Marlowe.ACTUS.TestFrameworkMarlowe
-        Spec.Marlowe.ACTUS.QCGenerator
-        Spec.Marlowe.ACTUS.QCTests
-    build-depends:
-        actus-core,
-        actus-haskell,
-        actus-testframework,
-        vector -any,
-        unordered-containers -any,
-        scientific -any,
-        extra -any,
-        utf8-string -any,
-        aeson -any,
-        base >=4.9 && <5,
-        containers -any,
-        hedgehog -any,
-        lens -any,
-        bytestring -any,
-        tasty -any,
-        tasty-hunit -any,
-        tasty-quickcheck -any,
-        text -any,
-        cborg,
-        plutus-ledger -any,
-        plutus-contract -any,
-        marlowe,
-        marlowe-actus,
-        plutus-tx -any,
-        template-haskell -any,
-        marlowe -any,
-        time -any,
-        validation -any,
-        pretty-simple -any,
-        mtl -any,
-        sort -any,
-        QuickCheck
+  default-language: Haskell2010
+  hs-source-dirs: test, actus-core/test
+  type: exitcode-stdio-1.0
+  main-is: Spec.hs
+  other-modules:
+    Spec.Actus.Examples
+    Spec.Actus.Haskell
+    Spec.Actus.Helper
+    Spec.Actus.TestFramework
+    Spec.Actus.TestFrameworkMarlowe
+    Spec.Actus.QCGenerator
+    Spec.Actus.QCTests
+  build-depends:
+    actus-core,
+    aeson -any,
+    base >=4.9 && <5,
+    bytestring -any,
+    containers -any,
+
+    scientific -any,
+    marlowe,
+    marlowe-actus,
+    mtl -any,
+    plutus-ledger -any,
+    plutus-contract -any,
+    plutus-tx -any,
+    pretty-simple -any,
+    sort -any,
+    tasty -any,
+    tasty-hunit -any,
+    tasty-quickcheck -any,
+    text -any,
+    time -any,
+    validation -any,
+    QuickCheck
   ghc-options:
     -Wall -Wnoncanonical-monad-instances
     -Wincomplete-uni-patterns -Wincomplete-record-updates
